backend/requirements.txt
Django==4.1.5
djangorestframework==3.14.0
drf-yasg==1.21.5
django-cors-headers==3.14.0
